儿童语音识别与判定Demo系统/backend/SpeechRec/requirements.txt
Django==3.0.6 -i https://pypi.tuna.tsinghua.edu.cn/simple
django-cors-headers
redis==2.10.6
django-redis
pytest-shutil
xlrd==1.2.0
Levenshtein
xpinyin
requests